requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1.7
django-active-link==0.1.8
django-extensions==3.2.1
django-import-export==3.2.0
et-xmlfile==1.1.0
graphviz==0.20.1
MarkupPy==1.14
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.1
Pillow==9.5.0
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
tzdata==2023.2
xlrd==2.0.1
xlwt==1.3.0
